pyproject.toml
[project]
name = "fullerton-ocr"
version = "0.1.0"
description = "OCR endpoint for medical document classification and extraction"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "google-generativeai>=0.8.0",
    "pymupdf>=1.24.0",
    "pdf2image>=1.17.0",
    "Pillow>=10.4.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
exclude = ["tests*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: tests that hit the live Gemini API (opt-in)",
]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
